Update object_detection/utils.py with object_detection/utils_new.py
